requirements.txt
# PyTorch ecosystem (com suporte CUDA)
torch==2.7.1
torchvision==0.22.1
#torchtext==0.18.0

# Machine Learning frameworks
scikit-learn==1.7.0
xgboost==3.0.2
lightgbm==4.6.0

# Deep Learning utilities
transformers==4.52.4
datasets==3.6.0
accelerate==1.7.0

# Computer Vision
opencv-python==4.11.0.86
pillow==11.2.1
albumentations==2.0.8
timm==1.0.15

# Data manipulation e análise
numpy==2.2.0
pandas==2.3.0
scipy==1.15.3

# Visualização
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.1.2
tensorboard==2.19.0
wandb==0.20.1

# Jupyter ecosystem
jupyter==1.1.1
jupyterlab==4.4.3
ipykernel==6.29.5
ipywidgets==8.1.7

# Utilitários gerais
tqdm==4.67.1
requests==2.32.4
PyYAML==6.0.2
python-dotenv==1.1.0

# Processamento de imagens avançado
imageio==2.37.0
scikit-image==0.25.2

# Métricas e avaliação
torchmetrics==1.7.2

# Otimização de hiperparâmetros
optuna==4.3.0

# Manipulação de dados em larga escala
dask==2025.5.1

# Monitoramento
psutil==7.0.0
GPUtil==1.4.0
pynvml==12.0.0